Changed the structure so that it is a proper python package.
I hadded different subpackages with modules grouped together within each one. I am preparing everything to be packaged in the first release. Eventually, I will remove the colors package. I do not add the dpg11 to the init since I have figured out how to properly structure everything. It now resides in the drivers subpackage.

Fixes Move the dpg11Device class to the function init file #7
