+++ dpg11_pylib/__init__.py
@@ -2,5 +2,4 @@ __version__ = '0.1.dev'
 
 
 
-def initial_test():
-    print("This is my print test")
+# from dpg11 import DPG11Device
