--- dpg11_pylib/baseFunctions.py
@@ -1,10 +1,0 @@
-"""
-The DPG11 has a number of functions that you can run by submitting this text file.
-This module will contain the functions to perform all of those functions by outputting a string. 
-I will update this in the future to have better functionality. But it should work as is for now. 
-"""
-
-# generates numpy arrays of waveforms that may be useful
-def ones_array(x, y):
-    # temp function for testing
-    return x + y
