+++ dpg11_pylib/driver/__init__.py
@@ -1,0 +1,1 @@
+from dpg11_pylib.driver.dpg11 import DPG11Device
