+++ dpg11_pylib/waveforms/__init__.py
@@ -1,0 +1,1 @@
+from dpg11_pylib.waveforms.waveforms import *
